% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc.R
\name{ancombc_tidy}
\alias{ancombc_tidy}
\title{Tidy ANCOM-BC output}
\usage{
ancombc_tidy(ancombc_out)
}
\arguments{
\item{ancombc_out}{output object from the ancombc() function}
}
\value{
a tibble of tidy data
}
\description{
Create a tidy table of ANCOM-BC output
}
