% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc.R
\name{ancombc_unbiased_abundances}
\alias{ancombc_unbiased_abundances}
\title{Get unbiased abundances from ANCOM-BC output}
\usage{
ancombc_unbiased_abundances(ancombc_out, phyloseq_obj)
}
\arguments{
\item{ancombc_out}{output object from the ancombc() function}

\item{phyloseq_ojb}{phyloseq object used as input for ancombc() function}
}
\value{
a data.frame of abundances
}
\description{
See https://bioconductor.org/packages/release/bioc/vignettes/ANCOMBC/inst/doc/ANCOMBC.html
for more info on unbiased abundances.
}
