% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{as.Num}
\alias{as.Num}
\title{convert to numeric while avoiding factor conversion issues}
\usage{
as.Num(x)
}
\arguments{
\item{x}{an iterable}
}
\value{
a numeric object
}
\description{
convert to numeric while avoiding factor conversion issues
}
