% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{beta2mtx}
\alias{beta2mtx}
\title{Convert tidy beta diversity table to a wide distance matrix}
\usage{
beta2mtx(dt)
}
\arguments{
\item{dt}{data.table, data.frame, or tibble}
}
\value{
a symmetric matrix of distances
}
\description{
The input should have the columns: Measure, SampleX, SampleY, Value
}
\details{
The output can be used for creating a PCoA. dendrogram, etc
}
