% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{calc_PCoA}
\alias{calc_PCoA}
\title{Wrapper for cmdscale}
\usage{
calc_PCoA(dist_mtx, k = 2)
}
\arguments{
\item{dist_mtx}{distance matrix object}
}
\value{
data.frame
}
\description{
Simple wrapper for cmdscale to provide data.frame formatted table.
If the distance matrices contain NAs, the samples containing NAs
will be removed (with a warning).
}
