% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{calculate_rarefaction_curves}
\alias{calculate_rarefaction_curves}
\title{Function for rarefaction analysis}
\usage{
calculate_rarefaction_curves(psdata, measures, depths, parallel = FALSE)
}
\arguments{
\item{psdata}{phyloseq object}

\item{measures}{Which diversity measures (see vegan package)}

\item{depths}{Which sequencing depths? Example: c(10, 100, 1000)}
}
\value{
A dataframe
}
\description{
Running estimate_richness_phy() at multiple subsampling depths
}
