% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{cat_file}
\alias{cat_file}
\title{pretty printing of a text file via cat}
\usage{
cat_file(file_name)
}
\arguments{
\item{file_name}{the name of the file to print}
}
\description{
This is most useful for working with IRkernl in Jupyter notebooks
}
