% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clustermq_logfile}
\alias{clustermq_logfile}
\title{Set a path for clustermq cluster job log files}
\usage{
clustermq_logfile(base_dir = "/ebio/abt3_scratch/")
}
\arguments{
\item{base_dir}{The base directory will the logfiles will be located.}
}
\value{
logfile path
}
\description{
Log files are optional for clustermq. The must be set in the template.
This function will create a unique directory within the "base_dir".
It will also return a path that you MUST use for the "log_file" parameter
in the Q template.
Moreover, the function will set the "clustermq.logfile" option to that
directory (used by clustermq_get_logs).
}
\details{
The function requires the uuid package.
}
\examples{
clustermq_setup()
tmpl = list(job_mem = '8G', log_file = clustermq_logfile())
fx = function(x, y) x * 2 + y
Q(fx, x=1:3, const=list(y=10), n_jobs=10, job_size=1, template=tmpl)
}
