% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clustermq_setup}
\alias{clustermq_setup}
\title{Set clustermq options}
\usage{
clustermq_setup(
  scheduler = c("sge", "multicore"),
  template = file.path(Sys.getenv("HOME"), ".clustermq.tmpl")
)
}
\arguments{
\item{scheduler}{The clustermq.scheduler option. Use "multicore" for local jobs.}

\item{template}{The clustermq.template option. It defaults to ~.clustermq.tmpl}
}
\description{
These options must be set before running clustermq
}
\examples{
clustermq_setup()              # sge job
clustermq_setup('multicore')   # local job
}
