% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{condaInfo}
\alias{condaInfo}
\title{"conda list" in R}
\usage{
condaInfo(conda_env)
}
\arguments{
\item{conda_env}{The name of the conda env to list}
}
\description{
This is most useful for working with IRkernl in Jupyter notebooks
}
