% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{df.dims}
\alias{df.dims}
\title{Changing number of rows/columns shown when printing a data frame}
\usage{
df.dims(nrows = 4, ncols = 20)
}
\arguments{
\item{nrows}{number of rows to print}

\item{ncols}{number of columns to print}
}
\description{
This is most useful for working with IRkernl in Jupyter notebooks
}
