% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dfhead}
\alias{dfhead}
\title{A simple dataframe summary}
\usage{
dfhead(df, n = 3)
}
\arguments{
\item{df}{dataframe object}

\item{n}{Number of lines to print}
}
\value{
a dataframe object
}
\description{
A simple dataframe summary
}
