% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dims}
\alias{dims}
\title{global change of plot size options}
\usage{
dims(w = 5, h = 5, res = 300)
}
\arguments{
\item{w}{figure width}

\item{h}{figure height}

\item{res}{figure resolution (DPI)}
}
\description{
This is most useful for working with IRkernl in Jupyter notebooks
}
