% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{dist_format}
\alias{dist_format}
\title{creating a string with distance & percent explained}
\usage{
dist_format(dist, PC1_perc_exp, PC2_perc_exp, label1 = 1, label2 = 2)
}
\arguments{
\item{dist}{str, distance metric}

\item{PC1_perc_exp}{float, percent variance explained for PC1}

\item{PC2_perc_exp}{float, percent variance explained for PC2}

\item{label1}{First PC label}

\item{label2}{Seconda PC label}
}
\value{
str, formatted as "metric, <PC1_perc_exp>%, <PC2_perc_exp>%"
}
\description{
creating a string with distance & percent explained
}
