% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{.HFE}
\alias{.HFE}
\title{supporting function for HFE}
\usage{
.HFE(
  brk,
  class_level,
  corr_cutoff = 0.5,
  freqCut = 99/1,
  uniqueCut = 1,
  quiet = TRUE
)
}
\value{
data.table
}
\description{
supporting function for HFE
}
