% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgnify_api.R
\name{.rec_search}
\alias{.rec_search}
\title{recursive search in a nested list; return flattened list}
\usage{
.rec_search(L, Lret = list(), full_path = c())
}
\value{
list
}
\description{
recursive search in a nested list; return flattened list
}
