% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{.well384_index}
\alias{.well384_index}
\title{making 384-well plate index}
\usage{
.well384_index()
}
\value{
named vector (Well --> location); column-wise location
}
\description{
making 384-well plate index
}
