% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{.well96_index}
\alias{.well96_index}
\title{making a 96-well plate index}
\usage{
.well96_index()
}
\value{
named vector (Well --> location); column-wise location
}
\description{
making a 96-well plate index
}
