% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena_api.R
\name{ena_get_filereport}
\alias{ena_get_filereport}
\title{Get file reports via ENA Portal API}
\usage{
ena_get_filereport(
  accession,
  fields = c("accession", "run_accession", "nominal_length", "read_count",
    "base_count", "library_selection", "environment_biome", "environment_feature",
    "environment_material", "instrument_model", "instrument_platform", "library_name",
    "library_strategy", "sample_accession", "sample_collection", "sampling_platform",
    "sequencing_method", "project_name"),
  base_url = "https://www.ebi.ac.uk/ena/portal/api/",
  ...
)
}
\arguments{
\item{accession}{Sample/run accession}

\item{fields}{Fields to return. Use ena_get_search_fields() to list all possible fields.}

\item{base_url}{ENA API base url}

\item{...}{Parameters passed to httr::GET}
}
\value{
data.frame
}
\description{
Works at least with sample and run accessions.
}
\details{
ENA Portal API: https://www.ebi.ac.uk/ena/portal/api/
}
\examples{
df = ena_get_filereport('ERR479486')
df = ena_get_filereport('SRS2472312')
}
