% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena_api.R
\name{ena_get_search_fields}
\alias{ena_get_search_fields}
\title{Get possible search fields}
\usage{
ena_get_search_fields(
  section = c("read_run", "study", "sample"),
  base_url = "https://www.ebi.ac.uk/ena/portal/api/",
  ...
)
}
\arguments{
\item{section}{Section to query (eg., read_run)}

\item{base_url}{ENA API base url}

\item{...}{Parameters passed to httr::GET}
}
\value{
data.frame
}
\description{
ENA Portal API: https://www.ebi.ac.uk/ena/portal/api/
}
\examples{
df = ena_get_search_fields()
df = ena_get_search_fields('sample')
}
