% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{estimate_rarified_richness}
\alias{estimate_rarified_richness}
\title{Helper Function for rarefaction analysis}
\usage{
estimate_rarified_richness(psdata, measures, depth)
}
\arguments{
\item{psdata}{phyloseq object}

\item{measures}{Which diversity measures}

\item{depth}{The sampling depth}
}
\value{
molten alpha diversity object
}
\description{
Helper Function for rarefaction analysis
}
