% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{estimate_richness_phy}
\alias{estimate_richness_phy}
\title{phyloseq::estimate_richness, but includes Faith's PD}
\usage{
estimate_richness_phy(physeq, split = TRUE, measures = NULL)
}
\arguments{
\item{physeq}{Phyloseq object}

\item{split}{Splitting the OTU table}

\item{measures}{Which diversity measures (Faith's PD = "FaithPD)}
}
\value{
Dataframe
}
\description{
See physeq::estimate richness for full details
}
