% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expand.grid.lower}
\alias{expand.grid.lower}
\title{expand.grid(), but just lower-triange comparisions}
\usage{
expand.grid.lower(x, y, diag = FALSE)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{diag}{include same-same comparisons ('a' <=> 'b')?}
}
\value{
a data.frame of all non-reciprical comparisons
}
\description{
This is useful when you want pairwise comparisons,
but you don't need the reciprical ('a' <=> 'b' & 'b' <=> 'a').
}
\examples{
expand.grid.lower(1:3, 1:3)
expand.grid.lower(1:3, 1:3, diag=TRUE)
}
