% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{extract_pltdt}
\alias{extract_pltdt}
\title{Extract data from ggplot object}
\usage{
extract_pltdt(plot_object, output_path)
}
\arguments{
\item{plot_object}{A ggplot object}

\item{output_path}{Where to write the output}
}
\description{
The data is written to files
}
