% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{fig_uuid}
\alias{fig_uuid}
\title{create UUID for figure file name}
\usage{
fig_uuid(full = FALSE)
}
\arguments{
\item{full}{Full length uuid or trimmed to just 24 char?}
}
\value{
character object
}
\description{
create UUID for figure file name
}
