% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{files_to_list}
\alias{files_to_list}
\title{convert a vector of file paths into a named list}
\usage{
files_to_list(files, label_index = -1)
}
\arguments{
\item{files}{Vector of file paths (eg., by using "list_files()")}

\item{label_index}{Which item in the path to return? 1-indexing. If <1, samples selected from the end.}
}
\value{
list of files
}
\description{
convert a vector of file paths into a named list
}
\examples{
 files = c('/path/to/project/Sample1/table.txt', '/path/to/project/Sample2/table.txt')
 files_to_list(files, -1)
 files = c('/path/to/project/Sample1.txt', '/path/to/project/Sample2.txt')
 files_to_list(files, 0)
}
