% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Fread}
\alias{Fread}
\title{Simple wrapper around data.table::fread}
\usage{
Fread(
  infile = NULL,
  cmd = NULL,
  sep = "\\t",
  check.names = TRUE,
  tmpdir = file.path("/ebio", "abt3_scratch", Sys.info()[["user"]], "R_tmp"),
  ...
)
}
\arguments{
\item{infile}{Input file name}

\item{cmd}{Command instead of input file (eg., "gunzip -c INFILE")}

\item{sep}{Value delimiter}

\item{check.names}{Format check column names}

\item{...}{Passed to data.table::fread}

\item{tmp_dir}{Temp file directory. Scratch directory by default}
}
\value{
data.table
}
\description{
Simple wrapper around data.table::fread
}
