% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{Init}
\alias{Init}
\title{Wrapper for calling other init functions}
\usage{
Init(nrows = 4, ncols = 20, threads = 4, scheduler = c("multicore", "sge"))
}
\arguments{
\item{nrows}{passed to df.dims()}

\item{ncols}{passed to df.dims()}

\item{threads}{passed to data.table::setDTthreads()}

\item{scheduler}{passed to clustermq_setup()}
}
\description{
Wrapper for calling other init functions
}
