% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_heatmap}
\alias{itol_heatmap}
\title{create itol heatmap file}
\usage{
itol_heatmap(
  df,
  dataset_label,
  out_file = "itol_heatmap.txt",
  out_dir = getwd(),
  tree = NULL,
  dist_method = "bray",
  color_scheme = c("color", "bw")
)
}
\arguments{
\item{df}{A data.frame, tibble, or data.table (see above)}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{tree}{Tree object used for ordering the heatmap columns; if NULL, the dist_method will be used to create the tree}

\item{dist_method}{vegan::vegdist method for creating the correlation dendrogram}

\item{color_scheme}{Heatmap color scheme. color = blue-orange-yellow; bw=white-grey-black}
}
\description{
https://itol.embl.de/help.cgi#heatmap
}
\details{
If a data.frame is provided, the row names must exactly match the tree node labels.
If a data.table or tibble is provided, the first column values must exactly match the tree node labels.
All other columns are assumed to be values for plotting.
}
\examples{
# data.frame
df = data.frame(sample1 = c(1,2,3), sample2 = c(3,2,1), row.names = LETTERS[1:3])
itol_heatmap(df, 'test_label')
# data.table
dt = data.table(taxa = LETTERS[1:3], sample1 = c(1,2,3), sample2 = c(3,2,1))
itol_heatmap(dt, 'test_label')
}
