% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_multibar}
\alias{itol_multibar}
\title{create itol multi-bar file}
\usage{
itol_multibar(
  df,
  dataset_label,
  out_file = "itol_multibar.txt",
  out_dir = getwd(),
  legend = NULL,
  WIDTH = 200,
  COLOR = "#ff0000"
)
}
\arguments{
\item{df}{A data.frame, tibble, or data.table (see above)}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{legend}{A list that includes shapes, colors, and labels (see \code{\link{itol_colorstrip}})}

\item{WIDTH}{Bar width}

\item{COLOR}{Legend color}
}
\description{
https://itol.embl.de/help.cgi#multibar
}
\details{
If a data.frame is provided, the row names must exactly match the tree node labels.
If a data.table or tibble is provided, the first column values must exactly match the tree node labels.
All other columns are assumed to be values for plotting.
}
\examples{
# data.frame
df = data.frame(value1 = c(1,2,3), value2 = c(3,2,1), row.names = LETTERS[1:3])
itol_multibar(df, 'test_label')
# data.table
dt = data.table(taxa = LETTERS[1:3], value1 = c(1,2,3), value2 = c(3,2,1))
itol_multibar(dt, 'test_label')
}
