% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_simplebar}
\alias{itol_simplebar}
\title{create itol simple-bar file
df = dataframe; the rownames should correspond with the tree labels}
\usage{
itol_simplebar(
  df,
  dataset_label,
  out_file,
  out_dir = NULL,
  legend = NULL,
  WIDTH = 200
)
}
\description{
create itol simple-bar file
df = dataframe; the rownames should correspond with the tree labels
}
