% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{list_files}
\alias{list_files}
\title{list.files with full.names=TRUE & recursive=TRUE}
\usage{
list_files(path, pattern = NULL, full.names = TRUE, recursive = TRUE, ...)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory,}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned.}

\item{full.names}{a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned}

\item{recursive}{logical. Should the listing recurse into directories?}
}
\value{
A character vector containing the names of the files in the specified directories
}
\description{
list.files with full.names=TRUE & recursive=TRUE
}
