% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{make_dir}
\alias{make_dir}
\title{A helper function for creating a directory (recursively)}
\usage{
make_dir(dir, quiet = FALSE)
}
\arguments{
\item{dir}{path for the new directory (will create recursively)}

\item{quite}{quite output}
}
\description{
A helper function for creating a directory (recursively)
}
