% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgnify_api.R
\name{mgnify_request}
\alias{mgnify_request}
\title{GET request from the ENA}
\usage{
mgnify_request(
  url,
  max_pages = NULL,
  query = list(),
  verbose = TRUE,
  cache_file = "mgnify_request.RDS",
  cache_break = 10,
  use_cache = TRUE,
  ...
)
}
\arguments{
\item{url}{ENA API url}

\item{max_pages}{Max number of pages to return. If NULL, all pages returned.}

\item{query}{Query list passed to httr::GET}

\item{verbose}{Verbose output?}

\item{cache_file}{File name to cache (checkpoint) the results. Useful for big queries in case the job is interuppted.}

\item{cache_break}{Write cache file every N pages.}

\item{use_cache}{Read the cache file, if it exists?}

\item{...}{Parameters passed to httr::GET}
}
\value{
data.frame
}
\description{
GET request from the ENA
}
\examples{
x = mgnify_request('https://www.ebi.ac.uk/metagenomics/api/v1/biomes', max_pages = 2)
}
