% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgnify_api.R
\name{mgnify_request_get}
\alias{mgnify_request_get}
\title{GET request to MGnify API}
\usage{
mgnify_request_get(url, page = 1, query = list(), verbose = TRUE, ...)
}
\arguments{
\item{url}{Query url}

\item{page}{Page of records to return}

\item{query}{List passed to httr::GET(query=)}

\item{verbose}{Verbose output?}
}
\value{
list(status = character, page = numberic, pages = numberic, data = data.frame)
}
\description{
GET request to MGnify API
}
\examples{
mgnify_request_get('https://www.ebi.ac.uk/metagenomics/api/v1/samples', query=list(instrument_platform = 'ILLUMINA'))
}
