% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml_tax_HFE}
\alias{ml_tax_HFE}
\title{Hierachical Feature Selection}
\usage{
ml_tax_HFE(brk, tax_level, corr_cutoff = 0.7, threads = 2, quiet = TRUE)
}
\arguments{
\item{brk}{data.table generated by read_bracken()}

\item{tax_level}{which taxonmoic level to use?}

\item{corr_cutoff}{features with >cutoff will be filtered to just one}
}
\value{
data.table of filtered features
}
\description{
For each clade (defined by tax_level), aggregate species abundances to genus, family, ..., to tax_level,
then filter out taxa that correlate strongly (just one selected)
}
