% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr_boruta_filter}
\alias{mlr_boruta_filter}
\title{Custom mlr filter for Boruta}
\usage{
mlr_boruta_filter()
}
\value{
Nothing, but "boruta.filter" filter will be registered
}
\description{
A custom mlr filter that uses Boruta to select important features
This function registers the "boruta.filter" filter to be used with
makeFilterWrapper and other mlr filter functions.
}
\details{
\itemize{
  \item target str; what is the target variable in the task object (default: 'Class')
  \item pValue float; see Boruta docs (default: 0.01)
  \item maxRuns int; see Boruta docs (default: 200)
  \item hostHistory bool; see Boruta docs (default: FALSE)
  \item withTentative bool; keep tentative features (default: TRUE)
  \item verbose bool; list features selected? (default: FALSE)
  \item mustKeep vector; features that cannot be filtered (default: NULL)
  \item threads int; number of threads to use for Boruta (default: 1)
}
}
