% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr_getNestedTuneResultsOptPathDf}
\alias{mlr_getNestedTuneResultsOptPathDf}
\title{Version of getNestedTuneResultsOptPathDf that actually works}
\usage{
mlr_getNestedTuneResultsOptPathDf(r, trafo = FALSE)
}
\arguments{
\item{r}{The result of resampling of a tuning wrapper}

\item{trafo}{Should the units of the hyperparameter path be converted to the transformed scale?}
}
\value{
data.frame
}
\description{
For main docs, see ?getNestedTuneResultsOptPathDf
}
