% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{p.dims}
\alias{p.dims}
\title{Global change of plot size options}
\usage{
p.dims(w = 5, h = 5, res = 200)
}
\arguments{
\item{w}{figure width}

\item{h}{figure height}

\item{res}{figure resolution (DPI)}
}
\description{
This is most useful for working with IRkernl in Jupyter notebooks
}
