% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{path_get_label}
\alias{path_get_label}
\title{splitting path and returning just one item in the vector}
\usage{
path_get_label(file_path, index)
}
\arguments{
\item{file_path}{File path}

\item{index}{Which item in the path to return? 1-indexing. If <1, samples selected from the end.}
}
\value{
string
}
\description{
splitting path and returning just one item in the vector
}
