% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{pcoa2df}
\alias{pcoa2df}
\title{Convert PCoA object to a tidy dataframe}
\usage{
pcoa2df(pcoa, k = 3)
}
\arguments{
\item{pcoa}{A pcoa object generated by cmdscale}

\item{k}{The number of PCs to keep}
}
\value{
A data.frame of PCoA points for the top k PCs
}
\description{
Convert PCoA object to a tidy dataframe
}
