% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phyloseq_rel_abund}
\alias{phyloseq_rel_abund}
\title{Transform abundances to relative}
\usage{
phyloseq_rel_abund(physeq_obj, percent_abund = TRUE)
}
\arguments{
\item{physeq_obj}{The phyloseq object}

\item{percent_abund}{Fractional or percent abundance?}
}
\value{
A phyloseq object
}
\description{
A simple wrapper for transform_sample_counts()
}
