% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{pipelineInfo}
\alias{pipelineInfo}
\title{pipeline sessionInfo}
\usage{
pipelineInfo(pipeline_path, head_n = 10)
}
\arguments{
\item{pipeline_path}{The path to the pipeline directory}

\item{head_n}{The number of lines to print from the readme}
}
\description{
sessionInfo for LeyLab snakemake pipelines
}
