% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot}
\alias{Plot}
\title{plot figure and save the figure grob object to a file at the same time}
\usage{
Plot(
  p,
  file = NULL,
  path = NULL,
  suffix = "",
  saveObj = TRUE,
  saveImg = FALSE,
  width = NA,
  height = NA,
  ...
)
}
\arguments{
\item{p}{Plot object (ggplot2, base, etc)}

\item{file}{File name to write. If NULL, the name will be based on the md5sum of the object, so the name will change if the object changes.}

\item{path}{Path to write to. If NULL, the path will be .figures/.}

\item{suffix}{File name suffix (eg., '.png')}

\item{saveObj}{Write the Robj to a file?}

\item{saveImg}{Write the image to a file?}

\item{width}{Figure width. If NA, uses global options}

\item{height}{Figure height. If NA, uses global options}
}
\description{
This is most useful for working with IRkernl in Jupyter notebooks
}
