% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_profiling.R
\name{read_bracken}
\alias{read_bracken}
\title{function for reading in bracken taxonomy files}
\usage{
read_bracken(F, to_remove = "_num")
}
\arguments{
\item{F}{Path to bracken table file}

\item{to_remove}{Remove either columns ending in "_num" or "_frac"}
}
\value{
tibble
}
\description{
Either columns ending in "_num" or "_frac" will be removed (just raw or fractional abundances).
The table will be converted to long form (sample ~ abundance).
Taxonomy will be split into separate levels
}
