% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readLinesTail}
\alias{readLinesTail}
\title{Read the last N lines of a file}
\usage{
readLinesTail(x, n, ...)
}
\arguments{
\item{x}{The file name}

\item{n}{The last N lines to read}

\item{...}{Passed to scan()}
}
\description{
Read the last N lines of a file
}
