% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Robj_md5sum}
\alias{Robj_md5sum}
\title{Dump an R object as text to a temp file and get the md5sum of the file}
\usage{
Robj_md5sum(Robj)
}
\arguments{
\item{Robj}{Any R object}
}
\value{
md5sum
}
\description{
Dump an R object as text to a temp file and get the md5sum of the file
}
