% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scale_fill_all}
\alias{scale_fill_all}
\title{Great a better coloring scheme for taxon abundance barcharts}
\usage{
scale_fill_all(..., return_hex = FALSE)
}
\arguments{
\item{...}{Parameters passed to scale_fill_manual()}

\item{return_hex}{Return a vector of color hexidecimals instead of a plotting object.}
}
\value{
ScaleContinuous/ggproto object or vector
}
\description{
The default coloring scheme for ggplot2 makes it hard to
distinguish among data points in complex bar charts (eg., taxa plots).
This function is a wrapper around scale_color_continuous()
which changes the color scheme used.
}
\examples{
ggplot(mpg, aes(fl, hwy, fill=model)) +
  geom_bar(stat='identity') +
  scale_fill_all()
}
