% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{send_email}
\alias{send_email}
\title{A helper function to send an email via the mail bash cmd}
\usage{
send_email(
  body,
  subject = "R job complete",
  email = NULL,
  email_ext = "tuebingen.mpg.de"
)
}
\arguments{
\item{body}{The email body}

\item{subject}{The email subject line}

\item{email}{The email address. If NULL, then username used}

\item{email_ext}{The part after the "at" symbol}
}
\value{
The output of the system() call
}
\description{
A helper function to send an email via the mail bash cmd
}
