% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{size_objects}
\alias{size_objects}
\title{Returns the sizes of R objects}
\usage{
size_objects(Robj)
}
\arguments{
\item{Robj}{Vector with the names of R objects as characters}
}
\value{
A list with the name of R objects as names and the formatted size of the objects
}
\description{
Returns the sizes of R objects
}
