% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{snakemakeInfo}
\alias{snakemakeInfo}
\title{snakemake conda info}
\usage{
snakemakeInfo(config_file, pipeline_dir, conda_env)
}
\arguments{
\item{config_file}{The path to the config file}

\item{pipeline_dir}{The path to the pipeline_directory}

\item{conda_env}{The conda env that has snakemake installed}
}
\value{
The environment info
}
\description{
snakemake conda info
}
