% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{split_path}
\alias{split_path}
\title{python's os.path.split() for R}
\usage{
split_path(x)
}
\arguments{
\item{x}{The full file path}
}
\value{
A vector of all path parts
}
\description{
python's os.path.split() for R
}
