% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{summary_x}
\alias{summary_x}
\title{Summary for numeric vectors that includes sd and stderr}
\usage{
summary_x(x, rnd = 3)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a matrix
}
\description{
sd = standard deviation
stderr = standard error of the mean (sd(x) / sqrt(length(x)))
}
