% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{taxonomy_levels}
\alias{taxonomy_levels}
\title{A simple function that returns a vector of taxonomy levels}
\usage{
taxonomy_levels()
}
\description{
This just saves some typing, since I find myself constantly typing out:
c('Domain', 'Phylum', 'Class', 'Order', 'Family', 'Genus', 'Species')
}
