% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{to_rds}
\alias{to_rds}
\title{Save object as RDS, with name automatically defined}
\usage{
to_rds(obj, file = NULL, path = NULL, suffix = "")
}
\arguments{
\item{file}{File name to write. If NULL, the name will be based on the md5sum of the object, so the name will change if the object changes.}

\item{path}{Path to write to. If NULL, the path will be .data/.}

\item{suffix}{File name suffix,}
}
\description{
Similar to the Plot() function, but for any R object.
This is useful for quickly saving data for use in other sessions.
For example, if one must compile tables of all p-values for manuscript submission.
}
