% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{unique_n}
\alias{unique_n}
\title{Pretty print number of unique elements in a vector}
\usage{
unique_n(x, label = "items")
}
\arguments{
\item{x}{a vector}

\item{label}{what do call the items in the vector (eg., "samples")}
}
\description{
The result will be cat'ed to the screen
}
