% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlab.R
\name{well2index}
\alias{well2index}
\title{Convert between wellID and column-num}
\usage{
well2index(x, plate_type = "96-well")
}
\arguments{
\item{x}{A vector of well IDs}

\item{plate_type}{Either 96-well or 384-well}
}
\value{
A vector of plate positions
}
\description{
Useful for converting between WellIDs (eg., "A2")
and well position in a plate (eg., 9)
}
