% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{write_table}
\alias{write_table}
\title{writing table convience function}
\usage{
write_table(df, file, ...)
}
\arguments{
\item{df}{Data.frame to write out}

\item{file}{Output file path}

\item{...}{Passed to write.table}
}
\description{
This is most useful for working with IRkernl in Jupyter notebooks
}
