% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{bash_job}
\alias{bash_job}
\title{bash job using conda env}
\usage{
bash_job(cmd, conda_env, stdout = TRUE, stderr = TRUE, print_output = TRUE)
}
\arguments{
\item{cmd}{The bash command in a string format}

\item{conda_env}{The conda env to use}

\item{stdout}{Print the stdout from the command?}

\item{stderr}{Print the stderr from the command?}

\item{quiet}{No printing}
}
\description{
The conda setup is assumed to be in your ~/.bashrc
If print_output == TRUE: the stdout/stderr will be printed instead of returned
Else: the stdout/stderr with be returned by the function
stderr/stdout is printed unless print_output==FALSE
}
