% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{bash_job}
\alias{bash_job}
\title{bash job using conda env}
\usage{
bash_job(cmd, conda_env, stdout = TRUE, stderr = TRUE)
}
\arguments{
\item{cmd}{bash command in a string format}

\item{conda_env}{the conda env to use}

\item{stdout}{print the stdout from the command?}

\item{stderr}{print the stderr from the command?}
}
\value{
stdout/stderr
}
\description{
The conda setup is assumed to be in your ~/.bashrc
}
