% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{calc_beta_div}
\alias{calc_beta_div}
\title{beta-diversity calculation}
\usage{
calc_beta_div(
  df,
  tree = NULL,
  method = c("wunifrac", "unifrac", "manhattan", "euclidean", "canberra", "clark",
    "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford",
    "raup", "binomial", "chao", "cao", "mahalanobis"),
  threads = 1
)
}
\arguments{
\item{df}{sample x taxon dataframe. Colnames (taxa) must match the tree tip labels if the tree is provided}

\item{tree}{phylogeny with tips matching the df colnames (only needed for wunifrac & unifrac methods)}

\item{method}{distance method (vegdist distances; wunifrac=Weighted Unifrac; unifrac=Unweighted Unifrac)}

\item{threads}{threads used for UniFrac calculations with rbiom}
}
\value{
data.frame
}
\description{
A wrapper around vegan::vegdist and rbiom (rbiom used for UniFrac calculations).
For unifrac: "wunifrac" = weighted unifrac, "unifrac" = unweighted unifrac.
The function returns a tidy dataframe of PCoA axes (PC1 & PC2),
percent variance explained for each PC.
}
\details{
Unifrac is calculated with the https://github.com/cmmr/rbiom package
(requires bioconductor packages).

If the goal is PCoA, then see the "tidy_PCoA" function.
}
