% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{.dist_fmt}
\alias{.dist_fmt}
\title{creating a string with distance & percent explained}
\usage{
.dist_fmt(dist, PC1_perc_exp, PC2_perc_exp)
}
\description{
creating a string with distance & percent explained
}
