% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{Fread}
\alias{Fread}
\title{Simple wrapper around data.table::fread}
\usage{
Fread(infile = NULL, cmd = NULL, sep = "\\t", check.names = TRUE, ...)
}
\arguments{
\item{infile}{input file name}

\item{cmd}{command instead of input file (eg., "gunzip -c INFILE")}

\item{sep}{value delimiter}

\item{check.names}{format check column names}

\item{...}{passed to data.table::fread}
}
\value{
data.table
}
\description{
Simple wrapper around data.table::fread
}
