% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_boxplot}
\alias{itol_boxplot}
\title{create itol boxplot file}
\usage{
itol_boxplot(
  df,
  dataset_label,
  out_file,
  out_dir = NULL,
  key_color = "#ff0000",
  WIDTH = 200
)
}
\arguments{
\item{df}{Dataframe, in which the rownames should correspond with the tree labels; the columns must specify: minimum,q1,median,q3,maximum,extreme_value1,extreme_value2}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{key_color}{The color for the legend key}

\item{WIDTH}{Maximum width}
}
\description{
https://itol.embl.de/help.cgi#boxplot
}
