% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_colorstrip}
\alias{itol_colorstrip}
\title{create itol colorstrip file}
\usage{
itol_colorstrip(df, dataset_label, out_file, out_dir = NULL, legend = NULL)
}
\arguments{
\item{df}{Dataframe, in which the rownames should correspond with the tree labels; the plotting parameter should be column 1}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{legend}{Specify particular legend}
}
\description{
create itol colorstrip file
}
