% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_externalshape}
\alias{itol_externalshape}
\title{create itol external shape file}
\usage{
itol_externalshape(
  df,
  dataset_label,
  out_file,
  out_dir = NULL,
  legend = NULL,
  WIDTH = 200
)
}
\arguments{
\item{df}{Dataframe, in which the rownames should correspond with the tree labels; other columns should be values corresponding to symbol size}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{legend}{Specify particular legend}
}
\description{
create itol external shape file
}
