% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_heatmap}
\alias{itol_heatmap}
\title{create itol heatmap file}
\usage{
itol_heatmap(
  df,
  dataset_label,
  out_file,
  out_dir = NULL,
  tree = NULL,
  dist_method = "bray",
  color_scheme = c("color", "bw")
)
}
\arguments{
\item{df}{Dataframe, in which the rownames should correspond with the tree labels; all columns should be numeric values for the heatmap}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{tree}{Tree object used for ordering the heatmap columns; if NULL, the dist_method will be used to create the tree}

\item{dist_method}{vegan::vegdist method for creating the correlation dendrogram}

\item{color_scheme}{Heatmap color scheme. color = blue-orange-yellow; bw=white-grey-black}
}
\description{
https://itol.embl.de/help.cgi#heatmap
}
