% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_multibar}
\alias{itol_multibar}
\title{create itol multi-bar file}
\usage{
itol_multibar(
  df,
  dataset_label,
  out_file,
  out_dir = NULL,
  legend = NULL,
  WIDTH = 200
)
}
\arguments{
\item{df}{Dataframe, in which the rownames should correspond with the tree labels}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{legend}{A list that includes shapes, colors, and labels (see \code{\link{itol_colorstrip}})}

\item{WIDTH}{Bar width}
}
\description{
https://itol.embl.de/help.cgi#multibar
}
