% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_simplebar}
\alias{itol_simplebar}
\title{create itol simple-bar file}
\usage{
itol_simplebar(
  df,
  dataset_label,
  out_file,
  out_dir = NULL,
  legend = NULL,
  WIDTH = 200
)
}
\arguments{
\item{df}{Dataframe, the rownames should correspond with the tree labels}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{legend}{Specify particular legend (see \code{\link{itol_colorstrip}})}

\item{WIDTH}{Bar width}
}
\description{
https://itol.embl.de/help.cgi#bar
}
