% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itol.R
\name{itol_symbol}
\alias{itol_symbol}
\title{create itol symbol file}
\usage{
itol_symbol(df, dataset_label, out_file, out_dir = NULL, MAXIMUM_SIZE = 50)
}
\arguments{
\item{df}{Dataframe, in which the rownames should correspond with the tree internal node labels, and other columns should be: symbol,size,color,fill,position,(label)}

\item{dataset_label}{What to label the itol dataset}

\item{out_file}{Name of the output file}

\item{out_dir}{Where to write the output}

\item{MAXIMUM_SIZE}{The max size of the symbols}
}
\description{
create itol symbol file
}
