% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml_tax_HFE}
\alias{ml_tax_HFE}
\title{Hierarchical Feature Selection}
\usage{
ml_tax_HFE(
  brk,
  tax_level,
  corr_cutoff = 0.7,
  threads = 2,
  freqCut = 95/1,
  uniqueCut = 5,
  quiet = TRUE
)
}
\arguments{
\item{brk}{data.table generated by read_bracken()}

\item{tax_level}{which taxonmoic level to use?}

\item{corr_cutoff}{features with >cutoff will be filtered to just one}

\item{freqCut}{as in caret::nearZeroVar; use NULL to skip}

\item{uniqueCut}{as in caret::nearZeroVar; use NULL to skip}
}
\value{
data.table of filtered features
}
\description{
For each clade (defined by tax_level), aggregate species abundances at each taxonomic
level up to the user-defined "tax_level", (optionally filter out near-zero features),
then filter out taxa that correlate strongly
(just one taxon is selected of those that correlate).
}
