% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{overlap}
\alias{overlap}
\title{Determine counts of setdiff, intersect, & union of 2 vectors}
\usage{
overlap(x, y)
}
\arguments{
\item{x}{vector1}

\item{y}{vector2}
}
\description{
The output is printed text of intersect, each-way setdiff, and union.
}
