% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{path_get_label}
\alias{path_get_label}
\title{splitting path and returning just one item in the vector}
\usage{
path_get_label(file_path, index)
}
\arguments{
\item{file_path}{File path(s). If vector or list of paths provided, then a list will be returned}

\item{index}{Which item in the path to return? 1-indexing. If <1, samples selected from the end. "O" will select the file name.}
}
\value{
string if 1 path, else list
}
\description{
This is useful for merging tables in which the individual table ID
is within the file path.
}
