% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_profiling.R
\name{read_bracken}
\alias{read_bracken}
\title{Function for reading in a bracken taxonomy table}
\usage{
read_bracken(
  infile,
  is_gzip = FALSE,
  n_lines = 0,
  keep_frac = TRUE,
  tax_levs = c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species")
)
}
\arguments{
\item{infile}{Path to bracken table file}

\item{is_gzip}{Is the table file gzip'ed?}

\item{n_lines}{Number of lines to read. If < 1, all lines will be read.}

\item{keep_frac}{If TRUE, keep all columns ending in "_frac"; otherwise, keep "_num" columns.}

\item{tax_levs}{Taxonomic levels to separate the taxonomy column into.}
}
\value{
data.table
}
\description{
The table will be converted to long form (sample ~ abundance).
Only "_frac" or "_num" columns will be kept (see "keep_frac").
Taxonomy will be split into separate levels (see "tax_levs").
tidytable (w/ data.table) used to speed the process up.
}
