% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{summary_x}
\alias{summary_x}
\title{Summary for numeric vectors that includes sd and stderr}
\usage{
summary_x(x, label = NULL, sel_col = NULL, rnd = 3)
}
\arguments{
\item{x}{a numeric vector}

\item{label}{row name label for the output. If NULL, then the label will be the input object label.}

\item{sel_col}{If "x" is data.table or data.frame, which column to assess?}

\item{rnd}{number of digits to round sd and stderr to}
}
\value{
a matrix
}
\description{
sd = standard deviation
stderr = standard error of the mean (sd(x) / sqrt(length(x)))
}
