% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{tidy_pcoa}
\alias{tidy_pcoa}
\title{PCoA on a 'long' (tidy) tibble, and a long tibble is returned}
\usage{
tidy_pcoa(
  df,
  taxon_col,
  sample_col,
  abundance_col,
  dists = c("bray", "jaccard"),
  tree = NULL,
  threads = 1
)
}
\arguments{
\item{df}{data.frame or tibble}

\item{taxon_col}{the column specifying taxa or OTUs (no quotes needed)}

\item{sample_col}{the column specifying sample names (no quotes needed)}

\item{abundance_col}{the column specifying the taxon abundances in each sample (no quotes needed)}

\item{dists}{vector of beta-diversity distances ('wunifrac' = weighted UniFrac, 'unifrac' = unweighted Unifrac, see vegan::vegsist for others)}

\item{tree}{phylogeny for UniFrac calculations. It can have more tips that what is in the data.frame}

\item{threads}{number of parallel calculations (1 thread per distance)}
}
\value{
data.frame
}
\description{
Perform PCoA in a "tidy" way.
If multiple methods provided, all PCoA results will be combined into one data.frame.
}
